refactor(#28): Change root java package and maven groupId
