+++ javalint-cli/src/main/java/io/github/kamilperczynski/javalint/cli/commands/CheckFormattingCommandEvents.kt
@@ -1,22 +1,20 @@
-package io.github.kamilperczynski.javalint.formatter.output
+package io.github.kamilperczynski.javalint.cli.commands
 
 import io.github.kamilperczynski.javalint.formatter.FormatterEvents
 import java.nio.file.Path
 
-class FixFormattingCommandEvents(private val homePath: Path) : FormatterEvents {
+class CheckFormattingCommandEvents(private val homePath: Path) : FormatterEvents {
 
   private var formattingStartedAt: Long = 0
 
-  private var filesWithCorrectStyleCount = 0
   var reformattedFilesCount = 0
     private set
 
   override fun formattingStarted() {
     print(ConsoleColor.WHITE_BOLD_BRIGHT)
-    print("Formatting all files at: $homePath")
+    print("Checking files at: $homePath against the code style")
     print(ConsoleColor.RESET)
     println()
-
   }
 
   override fun fileFormattingStarted(path: Path) {
@@ -29,7 +27,7 @@ class FixFormattingCommandEvents(private val homePath: Path) : FormatterEvents {
     print("".padStart(10))
     print(" (ignored)".padStart(10))
     print(ConsoleColor.RESET)
-    println()
+    print('\r')
   }
 
   override fun fileFormattingEnd(path: Path, isModified: Boolean) {
@@ -39,23 +37,27 @@ class FixFormattingCommandEvents(private val homePath: Path) : FormatterEvents {
       reformattedFilesCount++
       print(path.toString().padEnd(110))
       print(" ${duration}ms".padStart(10))
+      print("            ")
       println()
 
     } else {
-      filesWithCorrectStyleCount++
-
       print(ConsoleColor.BLACK_BRIGHT)
       print(path.toString().padEnd(110))
       print(" ${duration}ms".padStart(10))
       print(" (unchanged)")
       print(ConsoleColor.RESET)
-      println()
+      print('\r')
     }
   }
 
   override fun formattingEnd() {
-    println()
-    println("✅ Reformatted $reformattedFilesCount files, correctly formatted files: $filesWithCorrectStyleCount")
+    println(" ".repeat(132))
+
+    if (reformattedFilesCount > 0) {
+      println("⛔ Found $reformattedFilesCount files with incorrect formatting")
+    } else {
+      println("✅ All files matched the code style")
+    }
   }
 
 }
