+++ javalint-cli/src/main/java/io/github/kamilperczynski/javalint/cli/commands/FixFormattingCommandEvents.kt
@@ -1,20 +1,21 @@
-package io.github.kamilperczynski.javalint.formatter.output
+package io.github.kamilperczynski.javalint.cli.commands
 
 import io.github.kamilperczynski.javalint.formatter.FormatterEvents
 import java.nio.file.Path
 
-class CheckFormattingCommandEvents(private val homePath: Path) : FormatterEvents {
+class FixFormattingCommandEvents(private val homePath: Path) : FormatterEvents {
 
   private var formattingStartedAt: Long = 0
 
-  var reformattedFilesCount = 0
-    private set
+  private var filesWithCorrectStyleCount = 0
+  private var reformattedFilesCount = 0
 
   override fun formattingStarted() {
     print(ConsoleColor.WHITE_BOLD_BRIGHT)
-    print("Checking files at: $homePath against the code style")
+    print("Formatting all files at: $homePath")
     print(ConsoleColor.RESET)
     println()
+
   }
 
   override fun fileFormattingStarted(path: Path) {
@@ -27,7 +28,7 @@ class CheckFormattingCommandEvents(private val homePath: Path) : FormatterEvents
     print("".padStart(10))
     print(" (ignored)".padStart(10))
     print(ConsoleColor.RESET)
-    print('\r')
+    println()
   }
 
   override fun fileFormattingEnd(path: Path, isModified: Boolean) {
@@ -37,27 +38,23 @@ class CheckFormattingCommandEvents(private val homePath: Path) : FormatterEvents
       reformattedFilesCount++
       print(path.toString().padEnd(110))
       print(" ${duration}ms".padStart(10))
-      print("            ")
       println()
 
     } else {
+      filesWithCorrectStyleCount++
+
       print(ConsoleColor.BLACK_BRIGHT)
       print(path.toString().padEnd(110))
       print(" ${duration}ms".padStart(10))
       print(" (unchanged)")
       print(ConsoleColor.RESET)
-      print('\r')
+      println()
     }
   }
 
   override fun formattingEnd() {
-    println(" ".repeat(132))
-
-    if (reformattedFilesCount > 0) {
-      println("⛔ Found $reformattedFilesCount files with incorrect formatting")
-    } else {
-      println("✅ All files matched the code style")
-    }
+    println()
+    println("✅ Reformatted $reformattedFilesCount files, correctly formatted files: $filesWithCorrectStyleCount")
   }
 
 }
