+++ javalint-ec/src/main/resources/javalint-codegen.hbs
@@ -1,4 +1,4 @@
-package {{outputPackage}};
+package {{outputPackage}}
 
 import javax.annotation.processing.Generated
